Implemented handling of weights and biases for lstm.
Only thing missing now is handling of activations i.e. output and cell state needs to be distinguished. As noted in the code comment, it is probably best to minimize based on output as well as cell state as the cell state is recurrent on itself. If it was not implemented that way, minimizing only on the output would misrepresent fixed points of the lstm.