Implemented characterization of fixed points
Computation of Jacobian was compared for packages numdifftools and autograd. Although results were not the same, the were close within an absolut difference of 1e-08. It was decided to use numdifftools as it seemd less complicated to use with numpy. However, it was found on the interent that autograd might be more robust when it comes to finite difference approximations. Furthermore, it was first tried to implement fixed point classification rules from Strogatz: Nonlinear Dynamics and Chaos (2015). However, it is assumed that at least in simple fixed point problems, only two types of fixed points, namely stable and unstable ones would appear. Therefore, it is implemented for now that a fixed point is considered stable if all eigenvalues are smaller than 1. I.e. all modes converge to the fixed point. Otherwise, the stable modes(i.e. directions) of the fixed point are computed and can be displayed in a figure at a later point. Here the code is not working yet. It needs to be figured out how to communicate information through the parallelization to the final output and the plotting function.