Implemented parallel optimization for scipyfpf
apparently multiprocessing can't pickel fun objects of either functions or lambda statements. Thus as of this point the code is not working. A workaround would be to define a new function inside the optimization function. That would only involve pickling matrices, integers and strings. However, it would be stupid and potentially inefficient, as it would mean that the function would have to be reevaluated each time it is built. Therefore, I will also consider looking for other workarounds.