Attempted to adjust lstm implementation for adam
As indicated by the update rule for the lstm, a dynamical system based on an lstm would require two inputs, the hidden and the cell state. Such functionality has been implemented in such a way that it treats the hidden and the cell state as one fixed point while only separating the two when evaluating the function. However some things remain unclear. First, it may be that the final fixedpoint can only be represented by the hidden state. In that case, the fixedpointfinder would need some functionality to deliver on that. Second, tensorflow has functionality to return_state of the lstm. However, it will only return the full state at the end of each batch. As of now a workaround has been found by simply using many batches to have many samples to draw from. It does of course barely sample the true depth of the behavior. As it seems necessary and e.g. mattgolub uses an lstm statetuple object in his version of a fixed point finder, there has to be a way to get the full state from tensorflow that is not clear yet. On a sidenote: optimization routine as of now was working suprisingly well and the objective could be minimized quite nicely.