Updated fixedpoint classification
It is not easy to find information on classification in linear system with more than 2 dimensions. It was obvious that the previous classification did not work because it was classifying bisaddle points as stable points although they clearly had two unstable modes. I ended up following this forum in which a student mentioned to have studied under professor strogatz: https://www.physicsforums.com/threads/classification-of-fixed-points-of-n-dimensional-linear-dynamical-system.517582/. Classification is now binary. A fixed point can either have no unstable modes, i.e. a stable fixe point or one or more unstable modes (might get limited later on), i.e. a saddle point. It is purely decided on the sign of the eigenvalues of the jacobian matrix.