Some problems with minimization for lstm remaining
It still needs to be solved how the lstm can be minimized. In principle, minimization of the objective function works. However, the results remain to be a single point somewhere in the center of the block. As of now minimization is performed using both h and c states. The states are concatenated and used as input for two distinct functions. One function returns the c state and the other the h state. However, both functions depend on both states. The function results are themselves also concatenated so that the initial states can be substracted and fed into the rest of the objective. Note that this is what happens here https://github.com/mattgolub/fixed-point-finder/blob/df076121b57213323002bc4b6f04a24ed0784f6f/FixedPointFinder.py#L556 as well. The two states and their respective results are flattened into a vector valued function. Thus optimization will make use of all states and the jacobian is expected to have the size 2*n_units x 2*n_units. It seems the current implementation only a single fixed point can ever be found. It suggests that the equations remain uncorrect. Several things have indicated that. \n a. the velocities do not seem to be computed correctly. If simply evaluating velocities and subsequently plotting activations colored by velocities, each of the edges gets assigned a different velocity within a very narrow range from ca. 0.2-0.4. Suggesting that the equations do not represent how the activations were generated. A way to solve that would be to look at how it was originally implemented in keras. However that has been done multiple times now and did not provide any elusive insights yet. \n b. the equations as they are implemented atm seem to have some sort of convexity towards the middle of the block. Anywhere else one does not find a fixed point. This is irrespective of hps settings and sample size.