Moved fixedpointfinder and related functions to fixedpointfinder package
Files have been cleaned up a little. New Tffiedpointfinder class has been set up. It is not yet functioning though. Will be worked on in the future, also please note that the lstm is not yet correctly implemented for fixed point finding. Some changes have been made to threebitflipflop. The class now also has functionality to use pretrained models. Meaning it can incorporate pretrained recurrent_kernels and has functionality to train or not to train the recurrent_kernel. Therefore the two class SimpleRNN and SimpleRNNCell were inherited and manipulated. This may follow for other architectures in the future. Rnn construction will get its own class and package and there will be a split as of now.