Set up rdm across time
rsa class is intialized with activations. It will have many functions to compute rdms across a lot of dimensions and perspectives. First to be implemented is across time. The function allows to specify an interval across which the dataset will be windowed. based on that rdm will be computed. The mean is computed over all correlations in that time interval. A representational dissimilarity matrix is returned to be visualized.