First hint at relation of eigenvalues and vectors to system
When the comlex eigenvalue is transformed into polar coordinates and the angle of the imaginary part to the real part is computed, 17 nonzero angles can be found for the three bit flip flop system. Among those three have the exact value of 180 degree. The other 14 are pairs of 7 angles each with positive and negative direction. Furthermore, when comparing those angles with the angles of computed fixedpoints for the system, the angles of the fixed points can be described by one or a combination of angles derived from the complex eigenvalues. Thus this finding suggests that the system uses vectors corresponding to complex eigenvalues to perform rotations such that together with another set of eigenvectors, the system points towards the state that is intended to be memorized at a certain timeframe. Additionally, if one looks at the norm of the computed fixedpoints one can find that their distance from the origin can be described by a combination of the real valued eigenvalues that have no complex part. Therefore, this finding suggests that the system is using linear combinations of eigenvectors and rotations of those to define 26 unique memory states in which the system can be found.